Sensor begin fix (#5)
* added begin to sensors

* fixed wrong refresh logic on sensorpage

* uncommented sensor code

+++ lib/sensors/structs/sensorReadings.h
@@ -2,7 +2,7 @@
 
 struct SensorReadings
 {
-    uint16_t soilMoisture;
+    uint8_t soilMoisture;
     float airHumidity;
     float airTemp;
     float soilTemp;
